Fix scripts

+++ build.sh
@@ -2,10 +2,19 @@
 
 set -eo pipefail
 
-mod_dir=".release"
+repo_dir=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
+mod_dir="${repo_dir:?}/.release"
+linked_folder="${repo_dir:?}/.factorio-test-data/mods/crafting-combinator-tamj0rd2"
 
-rm -rf "${mod_dir}"
+rm -rf "${mod_dir:?}"/*
 npm run build
 cp mod_info.json "${mod_dir}/info.json"
+
+mkdir -p ".factorio-test-data/mods"
+
+#don't create a symlink if there's one already
+if [ ! -L "${linked_folder}" ] && [ ! -e "${linked_folder}" ]; then
+  ln -s "${mod_dir:?}" "${linked_folder:?}"
+fi
 
 echo "Built successfully"
